pyproject.toml
[project]
name = "natix"
version = "2.0.1"
description = ""
authors = [
    {name = "hayden-yuma",email = "hayden@yumaai.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"

[tool.poetry.dependencies]
# ^1.2.3 means >=1.2.3 and <2.0.0
# ~1.2.3 means >=1.2.3 and <1.3.0

torch = "^2.6.0"
torchvision = "^0.21.0"
torchaudio = "^2.6.0"
tensorflow = "^2.18.0"
tf-keras = "^2.18.0"
scikit-learn = "^1.5.2"

# Deep learning tools
transformers = "^4.47.0"
diffusers = "^0.32.2"
accelerate = "^1.2.0"
bitsandbytes = "0.46.1"
sentencepiece = "^0.2.0"
sentence-transformers = "^3.3.1"
timm = "^1.0.12"
einops = "^0.8.0"
ultralytics = "^8.3.44"

# Image/Video processing
datasets = "^3.1.0"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
scikit-image = "^0.24.0"

# Data and logging
wandb = "^0.19.0"
tensorboardx = "^2.6.2.2"
loguru = "^0.7.2"
httpx = "^0.28.0"
joblib = "^1.4.2"
bittensor = "^9.9.0"
h11 = "^0.16.0"
bittensor-cli = "^9.10.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pre-commit-hooks = "5.0.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"
pytest = "^8.3.3"
hypothesis = "^6.12.2"
flake8-pyproject = "~1.2.3"

[tool.black]
line-length = 130
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 130
skip_gitignore = true
skip = [
    ".venv",
    ".git",
    ".mypy_cache",
    "build",
    "dist",
]


[tool.flake8]
exclude = [
    # No need to traverse our git directory
    ".git",

    # There's no value in checking cache directories
    "__pycache__",

    # No need to traverse example code
    "*/examples/*",

    # No need to traverse docs
    "docs/*",

    # No need to traverse virtual environment
    ".venv/*",

    # No need to traverse these items
    ".mypy_cache/*",
    "build/*",
    "dist/*",
    ".vscode/*",
    "__init__.py"
]
max-complexity = 15
max-line-length = 130

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
